+++ app/Config/Routes.php
@@ -33,6 +33,7 @@ $routes->setAutoRoute(false);
 // We get a performance increase by specifying the default
 // route since we don't have to scan directories.
 $routes->resource('api/v1/users');
+$routes->resource('api/v1/groups');
 $routes->group('/', ['filter' => 'login'], function($routes){
 	/** Redirect attempts to Registration, Activation and Forgot/Resets */
 	$routes->match(['get', 'post'], 'register', 'CDR\Dashboard::index');